requirements.txt
# Dependencias estrictas de producción para nodo ARM64
numpy<2.0.0
aiogram==3.4.1
httpx==0.27.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
eth-account==0.11.0
APScheduler==3.10.4
cachetools==5.3.3
pydantic==2.6.4
python-dotenv==1.0.1
